Move update_body and skill rust related functions from player to Character (#37210)
